Finish main content and fix button
Add fourth and final block to the page.

Modify both index.html and style.css to include the last content which
consist of another button to sign up, as well as a little text at the
end of the page.
ALso fix dimensions of the "Sign up" button by removing the margin inside
the style for button in the css file.